+++ src/ServiceProvider.php
@@ -2,8 +2,6 @@
 
 namespace Butler\Health;
 
-use Composer\InstalledVersions;
-use Illuminate\Foundation\Console\AboutCommand;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
 
@@ -22,13 +20,5 @@ class ServiceProvider extends BaseServiceProvider
         if (! app()->routesAreCached() && config('butler.health.route')) {
             Route::get(config('butler.health.route'), Controller::class)->name('butler-health');
         }
-
-        AboutCommand::add('Environment', fn () => [
-            'Timezone' => config('app.timezone'),
-        ]);
-
-        AboutCommand::add('Butler Health', fn () => [
-            'Version' => ltrim(InstalledVersions::getPrettyVersion('glesys/butler-health'), 'v'),
-        ]);
     }
 }