+++ tests/RepositoryTest.php
@@ -9,15 +9,46 @@ class RepositoryTest extends AbstractTestCase
 {
     public function test_returns_correct_data()
     {
+        Repository::add('environment', ['foo' => 'bar']);
+        Repository::add('environment', fn () => ['foz' => 'baz']);
+
         $result = $this->travelTo(now(), function () {
             return (new Repository())();
         });
 
         AssertableJson::fromArray($result)
             ->has('about', fn (AssertableJson $json) => $json
-                ->where('environment.timezone', config('app.timezone'))
+                ->has('environment', fn (AssertableJson $json) => $json
+                    ->hasAll(
+                        'application_name',
+                        'laravel_version',
+                        'php_version',
+                        'composer_version',
+                        'environment',
+                        'debug_mode',
+                        'url',
+                    )
+                    ->where('timezone', config('app.timezone'))
+                    ->where('foo', 'bar')
+                    ->where('foz', 'baz')
+                )
+                ->has('cache', fn (AssertableJson $json) => $json
+                    ->hasAll('config', 'events', 'routes')
+                )
+                ->has('drivers', fn (AssertableJson $json) => $json
+                    ->hasAll(
+                        'broadcasting',
+                        'cache',
+                        'database',
+                        'logs',
+                        'mail',
+                        'octane',
+                        'queue',
+                        'session',
+                    )
+                )
                 ->has('butler_health.version')
-                ->etc())
+            )
             ->where('checks', [
                 [
                     'name' => 'Test Check',