+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Request info in response.
 
+### Changed
+- **BREAKING**: Snake case `runtimeInMilliseconds`.
+
 ## [0.5.1] - 2023-03-10
 ### Added
 - Uniqueness to custom "about" information.