+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Changed
 - **BREAKING**: Require Laravel 11.
 - **BREAKING**: Camel case keys in response.
+
+### Added
+- **BREAKING**: Enum `ResultState`.
 
 ## [0.5.2] - 2023-12-29
 ### Added