+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - **BREAKING**: Enum `ResultState`.
 - Config `butler.health.heartbeat.report`.
+- Config `butler.health.heartbeat.timeout`.
 
 ## [0.5.2] - 2023-12-29
 ### Added