+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.5.3] - 2024-08-01
 ### Added
 - Log driver for heartbeats.
-
 
 ## [0.5.2] - 2023-12-29
 ### Added